Rewrite of original code as a single Python script with parallelization of loops with multiprocessing.
